861m9x2bf: controls configurator in one component
