+++ front/src/schema/configurator/ClickedGunPart.ts
@@ -1,0 +1,6 @@
+export interface ClickedGunPart {
+    itemId: number;
+    productId: number;
+    parentId: number;
+    type: string;
+}
